requirements.txt
configparser
httpx
pydantic
python-dotenv
PyYAML